Initial commit
This is the AI I'm using (2016-10-25) for screeps. I managed to reach Top 10
from November 2015 - March 2016. Main Goal is to automated everything, no
manual interaction needed. Some attacking creeps can be steared via flags.

The AI automatically generated a layout for the room and builds the structures
for the current RCL. With a `scout` creep external harvest rooms are explored
and used. (SK rooms somwhow implemented, not enabled right now)
If the number of rooms are less than the GCL a new room
is acquired and build up. Also fallen rooms will be survived and basically
defended.
Some basic autoattacking is implemented, currently not enabled. Minerals
are fetched from the extractor and transported to the terminal. Reactions
are implemented (one reaction at a time, currently disabled). Depending on
a threshold minerals are sold on the market.