Replace spawn when enough energy is in storage
The spawn of the bot is placed randomly, to work proper the spawn
needs to be replaced. This happens on RCL 7 which is pretty late.
As soon as > 20k energy is in the storage four `planer` are build
the old spawn is destoryed and a correct placed one is build.