Fix issues in certain rooms
I think the underlying problem is the room setup, certain positions
(e.g. towerfiller) are initialized as `null`, `undefined` or
`-Infinity`.

These changes mails catch the errors in the methods which rely on a
proper rooms setup. So it is mostly a workaround.

Also a new `.eslint-basic.js` config is added to have basic linting to
improve on. The currently linting is pretty extensive and overwhelming.

CI server upgraded to nodejs 10.

Fixes #558
Fixes #554
Fixes #553