Improve carry handling and room startup
- In room startup carries call harvester on a lower energy/harvester
ratio
- Remove noise and improve log output in various places
- Remove restrictions for calling carries
- Simplify carry spawn logic: Use travel time + offset for the checking
interval, if dropped energy is greater than carry size, call a carry
- Replace some misplaced spawn logic, by simply make a random move when
a creep is stuck. The main problem is, when the spawn is on a pre
calculated path. When leaving the path `moveBackToPath` drops in and
should circumvent the misplaced spawn.
- On startup have only two harvester filling up the spawn and
extensions, the others continue upgrade the controller
- Harvesters don't build up roads, this led to harvester mainly busy
with moving between a road construction side in swamps and the spawn to
fill it up
- Refactor role carry to make the logic a bit clearer
- Harvester only execute the before storage logic. It was transferring
energy before, which led to filling up the spawn instead of upgrading
the controller on level 1
- Include structures in the test output

Fixes #496
Fixes #467
Fixes #464