Fix storage filler link position
The link with is next to filler needs to be the first in the position
list. Especially the link transfer logic relies on that fact.
Update bot version so that a new version is released for the private
server.
!! Updating the room layout version, this will recalculate the rooms and
might lead to a room rebuild. The only intended change is the position
of the storage link in `memory.position.structure.link`.

Fixes #582