Introduce `spawnIdle` (#711)
Store for each of `myRooms` the time a spawn is idle.
Include this in the reserver logic and restrict new reservation on a
spawnIdle threshold.

Currently too many rooms are tried to be reservered, while the actual
`reserver` creeps are queue due to busy spawns.
The `spawnIdle` check tackles this issue.