End of an epoch, friendly TooAngel NPC died (#712)
It is a sad day in the screeps multiverse. The friendly TooAngel NPC from the neighborhood died.
Since the official release of the game it was spending joy, some action via quests and helped improving defenses, while it was just trying to survive somehow for years.
Rest in peace friendly TooAngel NPC.

Some cleanup as condolences:
- Improve code quality
- Extract pixel generation
- Improve on reservations
- Reenable boost config
- Fix issues on mineral handling
- Split up quests in host and player
- Include attacked creeps in the reputation
- Start with autorespawner